This one only took me maybe a half hour to find.

+++ UIKit/UIScreen.h
@@ -17,7 +17,7 @@
 - (struct CGRect)applicationFrame;
 - (struct CGRect)bounds;
 #if 1 // iOS 4.x
-- (float) scale;
+- (CGFloat) scale;
 #endif
 
 @end
